added admin security promt and added env for clientside

+++ src/components/CommonListing/ProductsButtons/index.js
@@ -16,9 +16,20 @@ export default function ProductButton({ item }) {
   } = useContext(GlobalContext);
   const router = useRouter();
 
+  const pin = process.env.NEXT_PUBLIC_PIN_ADMIN;
+
   const isAdminView = pathName.includes("admin-view");
 
   async function handleDeleteProduct(item) {
+    const enteredPin = prompt("Enter Security Pin:");
+
+    if (enteredPin !== pin) {
+      toast.error("Only authenticated admin can make changes.", {
+        position: toast.POSITION.TOP_RIGHT,
+      });
+      return;
+    }
+
     setComponentLevelLoader({ loading: true, id: item._id });
 
     const res = await deleteAProduct(item._id);
@@ -41,26 +52,21 @@ export default function ProductButton({ item }) {
     <>
       <button
         onClick={() => {
+          const enteredPin = prompt("Enter Security Pin:");
+
+          if (enteredPin !== pin) {
+            toast.error("Only authenticated admin can make changes.", {
+              position: toast.POSITION.TOP_RIGHT,
+            });
+            return;
+          }
           setCurrentUpdatedProduct(item);
           router.push("/admin-view/add-product");
         }}
         className="mt-1.5 flex w-full justify-center bg-black px-5 py-3 text-xs font-medium uppercase tracking-wide text-white"
       >
         Edit
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          fill="none"
-          viewBox="0 0 24 24"
-          strokeWidth={1.5}
-          stroke="currentColor"
-          className="w-4 h-4 ml-3"
-        >
-          <path
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            d="m16.862 4.487 1.687-1.688a1.875 1.875 0 1 1 2.652 2.652L6.832 19.82a4.5 4.5 0 0 1-1.897 1.13l-2.685.8.8-2.685a4.5 4.5 0 0 1 1.13-1.897L16.863 4.487Zm0 0L19.5 7.125"
-          />
-        </svg>
+        {/* SVG icon */}
       </button>
       <button
         onClick={() => handleDeleteProduct(item)}
@@ -77,39 +83,13 @@ export default function ProductButton({ item }) {
         ) : (
           "Remove"
         )}
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          fill="none"
-          viewBox="0 0 24 24"
-          strokeWidth={1.5}
-          stroke="currentColor"
-          className="w-4 h-4 ml-3"
-        >
-          <path
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            d="m20.25 7.5-.625 10.632a2.25 2.25 0 0 1-2.247 2.118H6.622a2.25 2.25 0 0 1-2.247-2.118L3.75 7.5m6 4.125 2.25 2.25m0 0 2.25 2.25M12 13.875l2.25-2.25M12 13.875l-2.25 2.25M3.375 7.5h17.25c.621 0 1.125-.504 1.125-1.125v-1.5c0-.621-.504-1.125-1.125-1.125H3.375c-.621 0-1.125.504-1.125 1.125v1.5c0 .621.504 1.125 1.125 1.125Z"
-          />
-        </svg>
+        {/* SVG icon */}
       </button>
     </>
   ) : (
     <button className="mt-1.5 flex w-full justify-center bg-black px-5 py-3 text-xs font-medium uppercase tracking-wide text-white">
       Add To Cart
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        fill="none"
-        viewBox="0 0 24 24"
-        strokeWidth={1.5}
-        stroke="currentColor"
-        className="w-4 h-4 ml-3"
-      >
-        <path
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          d="M15.75 10.5V6a3.75 3.75 0 1 0-7.5 0v4.5m11.356-1.993 1.263 12c.07.665-.45 1.243-1.119 1.243H4.25a1.125 1.125 0 0 1-1.12-1.243l1.264-12A1.125 1.125 0 0 1 5.513 7.5h12.974c.576 0 1.059.435 1.119 1.007ZM8.625 10.5a.375.375 0 1 1-.75 0 .375.375 0 0 1 .75 0Zm7.5 0a.375.375 0 1 1-.75 0 .375.375 0 0 1 .75 0Z"
-        />
-      </svg>
+      {/* SVG icon */}
     </button>
   );
 }
